Moved the Metrics classes into their own packages instead of org.bstats
