+++ bstats-bungeecord/src/examples/java/ExamplePlugin.java
@@ -1,5 +1,5 @@
 import net.md_5.bungee.api.plugin.Plugin;
-import org.bstats.Metrics;
+import org.bstats.bungeecord.Metrics;
 
 import java.util.concurrent.Callable;
